@@ -1,39 +1,30 @@
+import 'package:bloc_poc/multibloc_provider_files.dart';
+import 'package:bloc_poc/screen/login_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:samplebloc/screen/login_page.dart';
-
-import 'multiprovider_config.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-
+  const MyApp({super.key});
   @override
   Widget build(BuildContext context) {
-    return MultiBlocProvider(
-      providers: multiProvider,
-      child: MaterialApp(
-          debugShowCheckedModeBanner: false,
-          theme: ThemeData(
-              appBarTheme: const AppBarTheme(
-                  elevation: 0,
-                  centerTitle: true,
-
-                  backgroundColor: Colors.teal,
-                  titleTextStyle: TextStyle(
-                      color: Colors.black,
-                      fontSize: 18,
-                      fontWeight: FontWeight.w600),
-                  iconTheme: IconThemeData(
+    return MaterialApp(
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+            appBarTheme: const AppBarTheme(
+                color: Colors.teal,
+                centerTitle: true,
+                elevation: 0,
+                titleTextStyle: TextStyle(
                     color: Colors.black,
-                    size: 30,
-                  )),
-             ),
-          home: const LoginPage()),
-    );
+                    fontSize: 18,
+                    fontWeight: FontWeight.w700))),
+        home: MultiBlocProvider(
+          providers: multiBlocProviderData,
+          child: const LoginPage(),
+        ));
   }
 }
